#275 FrostForge split to prefabs
